# server/swagger.yaml
openapi: 3.0.0
info:
  title: OneUp API
  description: API contract for the OneUp project.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /categories:
    post:
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Sample Category"
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  name:
                    type: string
                    example: "Sample Category"
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    get:
      summary: Get all categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 123
                    name:
                      type: string
                      example: "Sample Category"
                    created_at:
                      type: string
                      format: date-time
                      example: "2025-06-12T12:00:00Z"
                    updated_at:
                      type: string
                      format: date-time
                      example: "2025-06-12T12:00:00Z"
  /categories/{id}:
    get:
      summary: Get specific category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  name:
                    type: string
                    example: "Sample Category"
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    put:
      summary: Update specific category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 123
                name:
                  type: string
                  example: "Updated Category"
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  name:
                    type: string
                    example: "Updated Category"
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    delete:
      summary: Delete specific category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '204':
          description: Category deleted successfully
...